Apply stylecop fixes
